Create dylint/src/package_options/common.rs
