+++ dylint/src/package_options/mod.rs
@@ -18,6 +18,9 @@ use std::{
 use tempfile::tempdir;
 use walkdir::WalkDir;
 
+mod common;
+use common::parse_as_nightly;
+
 mod revs;
 use revs::Revs;
 
@@ -159,18 +162,4 @@ pub fn upgrade_package(opts: &opts::Dylint, upgrade_opts: &opts::Upgrade) -> Res
         .with_context(|| "Could not disable `Cargo.toml` backup")?;
 
     Ok(())
-}
-
-fn parse_as_nightly(channel: &str) -> Option<[u32; 3]> {
-    channel.strip_prefix("nightly-").and_then(parse_date)
-}
-
-fn parse_date(date_str: &str) -> Option<[u32; 3]> {
-    date_str
-        .split('-')
-        .map(str::parse::<u32>)
-        .map(Result::ok)
-        .collect::<Option<Vec<_>>>()
-        .map(<[u32; 3]>::try_from)
-        .and_then(Result::ok)
 }
